Remove find() method AR, move findOne(), FindAll(), FindBySql() to AQ.
