+++ src/ActiveQueryInterface.php
@@ -71,12 +71,14 @@ interface ActiveQueryInterface extends QueryInterface
      * The following are some usage examples:
      *
      * ```php
+     * // Create active query
+     * CustomerQuery = new ActiveQuery(Customer::class, $this->db);
      * // find customers together with their orders and country
-     * Customer::find()->with('orders', 'country')->all();
+     * CustomerQuery->with('orders', 'country')->all();
      * // find customers together with their orders and the orders' shipping address
-     * Customer::find()->with('orders.address')->all();
+     * CustomerQuery->with('orders.address')->all();
      * // find customers together with their country and orders of status 1
-     * Customer::find()->with([
+     * CustomerQuery->with([
      *     'orders' => function (ActiveQuery $query) {
      *         $query->andWhere('status = 1');
      *     },
@@ -113,4 +115,162 @@ interface ActiveQueryInterface extends QueryInterface
      * @return mixed the related record(s).
      */
     public function findFor(string $name, ActiveRecordInterface $model);
+
+    /**
+     * Returns a single active record active record instance by a primary key or an array of column values.
+     *
+     * The method accepts:
+     *
+     *  - a scalar value (integer or string): query by a single primary key value and return the corresponding record
+     *    (or `null` if not found).
+     *  - a non-associative array: query by a list of primary key values and return the first record (or `null` if not
+     *    found).
+     *  - an associative array of name-value pairs: query by a set of attribute values and return a single record
+     *    matching all of them (or `null` if not found). Note that `['id' => 1, 2]` is treated as a non-associative
+     *    array.
+     *
+     * Column names are limited to current records table columns for SQL DBMS, or filtered otherwise to be limited to
+     * simple filter conditions.
+     *
+     * That this method will automatically call the `one()` method and return an
+     * {@see ActiveRecordInterface|ActiveRecord} instance.
+     *
+     * > Note: As this is a short-hand method only, using more complex conditions, like ['!=', 'id', 1] will not work.
+     * > If you need to specify more complex conditions, in combination with {@see ActiveQuery::where()|where()} instead.
+     *
+     * See the following code for usage examples:
+     *
+     * ```php
+     * // find a single customer whose primary key value is 10
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->findOne(10);
+     *
+     * // the above code is equivalent to:
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->where(['id' => 10])->one();
+     *
+     * // find the customers whose primary key value is 10, 11 or 12.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->findOne([10, 11, 12]);
+     *
+     * // the above code is equivalent to:
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->where(['id' => [10, 11, 12]])->one();
+     *
+     * // find the first customer whose age is 30 and whose status is 1
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->findOne(['age' => 30, 'status' => 1]);
+     *
+     * // the above code is equivalent to:
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $query = $customerQuery->where(['age' => 30, 'status' => 1])->one();
+     * ```
+     *
+     * If you need to pass user input to this method, make sure the input value is scalar or in case of array condition,
+     * make sure the array structure can not be changed from the outside:
+     *
+     * ```php
+     * public function actionView(ServerRequestInterface $request)
+     * {
+     *     $id = (string) $request->getAttribute('id');
+     *
+     *     $aqClass = new ActiveQuery(Post::class, $db);
+     *     $query = $aqClass->findOne($id);
+     * }
+     *
+     * - Explicitly specifying the column to search, passing a scalar or array here will always result in finding a
+     * single record:
+     * $aqClass = new ActiveQuery(Post::class, $db);
+     * $query = $aqClass->findOne(['id' => $id);
+     *
+     * Do NOT use the following code! it is possible to inject an array condition to filter by arbitrary column
+     * values!:
+     * $aqClass = new ActiveQuery(Post::class, $db);
+     * $query = $aqClass->findOne($id);
+     * ```
+     *
+     * @param mixed $condition primary key value or a set of column values.
+     *
+     * @return ActiveRecordInterface|null instance matching the condition, or `null` if nothing matches.
+     */
+    public function findOne($condition): ?ActiveRecordInterface;
+
+    /**
+     * Returns a list of active record that match the specified primary key value(s) or a set of column values.
+     *
+     * The method accepts:
+     *
+     *  - a scalar value (integer or string): query by a single primary key value and return an array containing the
+     *    corresponding record (or an empty array if not found).
+     *  - a non-associative array: query by a list of primary key values and return the corresponding records (or an
+     *    empty array if none was found).
+     *    Note that an empty condition will result in an empty result as it will be interpreted as a search for
+     *    primary keys and not an empty `WHERE` condition.
+     *  - an associative array of name-value pairs: query by a set of attribute values and return an array of records
+     *    matching all of them (or an empty array if none was found). Note that `['id' => 1, 2]` is treated as
+     *    a non-associative array.
+     *    Column names are limited to current records table columns for SQL DBMS, or filtered otherwise to be limted to
+     *    simple filter conditions.
+     *
+     * This method will automatically call the `all()` method and return an array of
+     * {@see ActiveRecordInterface|ActiveRecord} instances.
+     *
+     * > Note: As this is a short-hand method only, using more complex conditions, like ['!=', 'id', 1] will not work.
+     * > If you need to specify more complex conditions, in combination with {@see ActiveQuery::where()|where()} instead.
+     *
+     * See the following code for usage examples:
+     *
+     * ```php
+     * // find the customers whose primary key value is 10.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->findAll(10);
+     *
+     * // the above code is equivalent to.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->where(['id' => 10])->all();
+     *
+     * // find the customers whose primary key value is 10, 11 or 12.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->findAll([10, 11, 12]);
+     *
+     * // the above code is equivalent to,
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->where(['id' => [10, 11, 12]])->all();
+     *
+     * // find customers whose age is 30 and whose status is 1.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->findAll(['age' => 30, 'status' => 1]);
+     *
+     * // the above code is equivalent to.
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customers = $customerQuery->where(['age' => 30, 'status' => 1])->all();
+     * ```
+     *
+     * If you need to pass user input to this method, make sure the input value is scalar or in case of array condition,
+     * make sure the array structure can not be changed from the outside:
+     *
+     * ```php
+     * public function actionView(ServerRequestInterface $request)
+     * {
+     *     $id = (string) $request->getAttribute('id');
+     *
+     *     $aqClass = new ActiveQuery(Post::class, $db);
+     *     $query = $aqClass->findOne($id);
+     * }
+     *
+     * Explicitly specifying the column to search, passing a scalar or array here will always result in finding a single
+     * record:
+     * $aqClass = new ActiveQuery(Post::class, $db);
+     * $aqCLass = $aqClass->findOne(['id' => $id]);
+     *
+     * Do NOT use the following code! it is possible to inject an array condition to filter by arbitrary column values!:
+     * $aqClass = new ActiveQuery(Post::class, $db);
+     * $aqClass = $aqClass->findOne($id);
+     * ```
+     *
+     * @param mixed $condition primary key value or a set of column values.
+     *
+     * @return array an array of ActiveRecord instance, or an empty array if nothing matches.
+     */
+    public function findAll($condition): array;
 }
