+++ src/ActiveQueryTrait.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\ActiveRecord;
 
 use ReflectionException;
+use Throwable;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
@@ -26,9 +27,9 @@ trait ActiveQueryTrait
      *
      * @param bool $value whether to return the query results in terms of arrays instead of Active Records.
      *
-     * @return ActiveQueryInterface the query object itself.
+     * @return $this the query object itself.
      */
-    public function asArray(?bool $value = true): ActiveQueryInterface
+    public function asArray(?bool $value = true): self
     {
         $this->asArray = $value;
 
@@ -50,12 +51,14 @@ trait ActiveQueryTrait
      * The following are some usage examples:
      *
      * ```php
+     * // Create active query
+     * CustomerQuery = new ActiveQuery(Customer::class, $db);
      * // find customers together with their orders and country
-     * Customer::find()->with('orders', 'country')->all();
+     * CustomerQuery->with('orders', 'country')->all();
      * // find customers together with their orders and the orders' shipping address
-     * Customer::find()->with('orders.address')->all();
+     * CustomerQuery->with('orders.address')->all();
      * // find customers together with their country and orders of status 1
-     * Customer::find()->with([
+     * CustomerQuery->with([
      *     'orders' => function (ActiveQuery $query) {
      *         $query->andWhere('status = 1');
      *     },
@@ -68,8 +71,8 @@ trait ActiveQueryTrait
      * For example, the following two statements are equivalent:
      *
      * ```php
-     * Customer::find()->with('orders', 'country')->all();
-     * Customer::find()->with('orders')->with('country')->all();
+     * CustomerQuery->with('orders', 'country')->all();
+     * CustomerQuery->with('orders')->with('country')->all();
      * ```
      * @param array|string $with
      *
@@ -131,11 +134,8 @@ trait ActiveQueryTrait
      * for details about specifying this parameter.
      * @param ActiveRecord[]|array $models the primary models (can be either AR instances or arrays)
      *
-     * @throws Exception
-     * @throws InvalidArgumentException
-     * @throws InvalidConfigException
-     * @throws NotSupportedException
-     * @throws ReflectionException
+     * @throws Exception|InvalidArgumentException|InvalidConfigException|NotSupportedException|ReflectionException
+     * @throws Throwable
      */
     public function findWith(array $with, array &$models): void
     {
@@ -158,11 +158,8 @@ trait ActiveQueryTrait
     }
 
     /**
-     * @param ActiveRecord $model
+     * @param ActiveRecordInterface $model
      * @param array $with
-     *
-     * @throws ReflectionException
-     * @throws InvalidArgumentException
      *
      * @return ActiveQuery[]|array
      */
