+++ src/ActiveRecordInterface.php
@@ -75,7 +75,7 @@ interface ActiveRecordInterface
      * Returns the old primary key value(s).
      *
      * This refers to the primary key value that is populated into the record after executing a find method (e.g.
-     * find(), findOne()).
+     * findOne()).
      *
      * The value remains unchanged even if the primary key attribute is manually assigned with a different value.
      *
@@ -97,222 +97,6 @@ interface ActiveRecordInterface
      * @return bool whether the given set of attributes represents the primary key for this active record.
      */
     public function isPrimaryKey(array $keys): bool;
-
-    /**
-     * Creates an {@see ActiveQueryInterface} instance for query purpose.
-     *
-     * The returned {@see ActiveQueryInterface} instance can be further customized by calling  methods defined in
-     * {@see ActiveQueryInterface} before `one()` or `all()` is called to return populated ActiveRecord instances.
-     *
-     * For example,
-     *
-     * ```php
-     * // find the customer whose ID is 1
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where(['id' => 1])->one();
-     *
-     * // find all active customers and order them by their age:
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where(['status' => 1])->orderBy('age')->all();
-     * ```
-     *
-     * This method is also called by {@see BaseActiveRecord::hasOne()} and {@see BaseActiveRecord::hasMany()} to create
-     * a relational query.
-     *
-     * You may override this method to return a customized query. For example,
-     *
-     * ```php
-     * class Customer extends ActiveRecord
-     * {
-     *     public function find()
-     *     {
-     *         // use CustomerQuery instead of the default ActiveQuery
-     *         return new CustomerQuery(get_called_class(), $db);
-     *     }
-     * }
-     * ```
-     *
-     * The following code shows how to apply a default condition for all queries:
-     *
-     * ```php
-     * class Customer extends ActiveRecord
-     * {
-     *     public function find()
-     *     {
-     *         return parent::find()->where(['deleted' => false]);
-     *     }
-     * }
-     *
-     * // Use andWhere()/orWhere() to apply the default condition
-     * // SELECT FROM customer WHERE `deleted`=:deleted AND age>30
-     * $customer = new Customer($db);
-     * $query = $customer->find()->andWhere('age>30')->all();
-     *
-     * // Use where() to ignore the default condition
-     * // SELECT FROM customer WHERE age>30
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where('age>30')->all();
-     *
-     * @return ActiveQueryInterface the newly created {@see ActiveQueryInterface} instance.
-     */
-    public function find(): ActiveQueryInterface;
-
-    /**
-     * Returns a single active record active record instance by a primary key or an array of column values.
-     *
-     * The method accepts:
-     *
-     *  - a scalar value (integer or string): query by a single primary key value and return the corresponding record
-     *    (or `null` if not found).
-     *  - a non-associative array: query by a list of primary key values and return the first record (or `null` if not
-     *    found).
-     *  - an associative array of name-value pairs: query by a set of attribute values and return a single record
-     *    matching all of them (or `null` if not found). Note that `['id' => 1, 2]` is treated as a non-associative
-     *    array.
-     *
-     * Column names are limited to current records table columns for SQL DBMS, or filtered otherwise to be limited to
-     * simple filter conditions.
-     *
-     * That this method will automatically call the `one()` method and return an
-     * {@see ActiveRecordInterface|ActiveRecord} instance.
-     *
-     * > Note: As this is a short-hand method only, using more complex conditions, like ['!=', 'id', 1] will not work.
-     * > If you need to specify more complex conditions, use {@see find()} in combination with
-     * {@see ActiveQuery::where()|where()} instead.
-     *
-     * See the following code for usage examples:
-     *
-     * ```php
-     * // find a single customer whose primary key value is 10
-     * $customer = new Customer($db);
-     * $query = $customer->findOne(10);
-     *
-     * // the above code is equivalent to:
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where(['id' => 10])->one();
-     *
-     * // find the customers whose primary key value is 10, 11 or 12.
-     * $customer = new Customer($db);
-     * $query = $customer->findOne([10, 11, 12]);
-     *
-     * // the above code is equivalent to:
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where(['id' => [10, 11, 12]])->one();
-     *
-     * // find the first customer whose age is 30 and whose status is 1
-     * $customer = new Customer($db);
-     * $query = $customer->findOne(['age' => 30, 'status' => 1]);
-     *
-     * // the above code is equivalent to:
-     * $customer = new Customer($db);
-     * $query = $customer->find()->where(['age' => 30, 'status' => 1])->one();
-     * ```
-     *
-     * If you need to pass user input to this method, make sure the input value is scalar or in case of array condition,
-     * make sure the array structure can not be changed from the outside:
-     *
-     * ```php
-     * public function actionView($id)
-     * {
-     *     $arClass = new Post($db);
-     *     $query = $arClass->findOne($id);
-     * }
-     *
-     * - Explicitly specifying the column to search, passing a scalar or array here will always result in finding a
-     * single record:
-     * $arClass = new Post($db);
-     * $query = $arClass->findOne(['id' => $id);
-     *
-     * Do NOT use the following code! it is possible to inject an array condition to filter by arbitrary column
-     * values!:
-     * $arClass = new Post($db);
-     * $query = $arClass->findOne($id);
-     * ```
-     *
-     * @param mixed $condition primary key value or a set of column values.
-     *
-     * @return ActiveRecordInterface|null instance matching the condition, or `null` if nothing matches.
-     */
-    public function findOne($condition): ?ActiveRecordInterface;
-
-    /**
-     * Returns a list of active record that match the specified primary key value(s) or a set of column values.
-     *
-     * The method accepts:
-     *
-     *  - a scalar value (integer or string): query by a single primary key value and return an array containing the
-     *    corresponding record (or an empty array if not found).
-     *  - a non-associative array: query by a list of primary key values and return the corresponding records (or an
-     *    empty array if none was found).
-     *    Note that an empty condition will result in an empty result as it will be interpreted as a search for
-     *    primary keys and not an empty `WHERE` condition.
-     *  - an associative array of name-value pairs: query by a set of attribute values and return an array of records
-     *    matching all of them (or an empty array if none was found). Note that `['id' => 1, 2]` is treated as
-     *    a non-associative array.
-     *    Column names are limited to current records table columns for SQL DBMS, or filtered otherwise to be limted to
-     *    simple filter conditions.
-     *
-     * This method will automatically call the `all()` method and return an array of
-     * {@see ActiveRecordInterface|ActiveRecord} instances.
-     *
-     * > Note: As this is a short-hand method only, using more complex conditions, like ['!=', 'id', 1] will not work.
-     * > If you need to specify more complex conditions, use {@see find()} in combination with
-     * > {@see ActiveQuery::where()|where()} instead.
-     *
-     * See the following code for usage examples:
-     *
-     * ```php
-     * // find the customers whose primary key value is 10.
-     * $customer = new Customer($db);
-     * $customers = $customer->findAll(10);
-     *
-     * // the above code is equivalent to.
-     * $customer = new Customer($db);
-     * $customers = $customer->find()->where(['id' => 10])->all();
-     *
-     * // find the customers whose primary key value is 10, 11 or 12.
-     * $customer = new Customer($db);
-     * $customers = $customer->findAll([10, 11, 12]);
-     *
-     * // the above code is equivalent to,
-     * $customer = new Customer($db);
-     * $customers = $customer->find()->where(['id' => [10, 11, 12]])->all();
-     *
-     * // find customers whose age is 30 and whose status is 1.
-     * $customer = new Customer($db);
-     * $customers = $customer->findAll(['age' => 30, 'status' => 1]);
-     *
-     * // the above code is equivalent to.
-     * $customer = new Customer($db);
-     * $customers = $customer->find()->where(['age' => 30, 'status' => 1])->all();
-     * ```
-     *
-     * If you need to pass user input to this method, make sure the input value is scalar or in case of array condition,
-     * make sure the array structure can not be changed from the outside:
-     *
-     * ```php
-     * public function actionView($id)
-     * {
-     *     $arClass = new Post($db);
-     *     $arClass = $arClass->findOne($id);
-     *     // ...
-     * }
-     *
-     * Explicitly specifying the colum to search, passing a scalar or array here will always result in finding a single
-     * record:
-     * $arClass = new Post($db);
-     * $arCLass = $arClass->findOne(['id' => $id]);
-     *
-     * Do NOT use the following code! it is possible to inject an array condition to filter by arbitrary column values!:
-     * $arClass = new Post($db);
-     * $arClass = $arClass->findOne($id);
-     * ```
-     *
-     * @param mixed $condition primary key value or a set of column values.
-     *
-     * @return array an array of ActiveRecord instance, or an empty array if nothing matches.
-     */
-    public function findAll($condition): array;
 
     /**
      * Updates records using the provided attribute values and conditions.
@@ -402,8 +186,8 @@ interface ActiveRecordInterface
      * Usage example:
      *
      * ```php
-     * $customer = new Customer($db);
-     * $customer = $customer->findOne($id);
+     * $customerQuery = new ActiveQuery(Customer::class, $db);
+     * $customer = $customerQuery->findOne($id);
      * $customer->name = $name;
      * $customer->email = $email;
      * $customer->update();
